requirements.txt
pre-commit
pytest
kivy
kivymd
pylint
pydocstyle
kivy_garden.graph
